make the users to be abe to login and see only their data

+++ src/firestore/codeItem/deleteCodeItem.ts
@@ -2,9 +2,15 @@
 
 import { db } from "@/firebase";
 import { deleteDoc, doc } from "firebase/firestore";
+import { findUserApi } from "../user/findUser";
 
-export const deleteCodeItem = async (docId: string): Promise<any> => {
+export const deleteCodeItemApi = async (
+  userId: string,
+  docId: string
+): Promise<any> => {
   try {
+    if (!userId) throw new Error("id of user not defiend");
+    await findUserApi(userId);
     const docRef = doc(db, "codeItems", docId);
     await deleteDoc(docRef);
     console.log(`Document with ID ${docId} deleted successfully.`);
